fix(release): Embed correct modules version
