+++ plugins/source/test/.gitignore
@@ -1,2 +1,2 @@
-cq-source-test
+test
 dist
