+++ plugins/source/test/main.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"github.com/cloudquery/cq-source-test/plugin"
+	"github.com/cloudquery/cloudquery/plugins/source/test/resources/plugin"
 	"github.com/cloudquery/plugin-sdk/serve"
 )
